1st version of API for querying node usage

+++ cmd/cluster_consolidated_node_cache.go
@@ -17,15 +17,16 @@ type NodeUsageDb struct {
 }
 
 
-func NewNodeUsageDB(path string) (*NodeUsageDb, error) {
+func NewNodeUsageDB(path string, createIfNotExist bool) (*NodeUsageDb, error) {
 	_, err := os.Stat(path)
-	if os.IsNotExist(err) {
+	if os.IsNotExist(err)  && createIfNotExist {
 		file, err := os.Create(path)
 		if err != nil {
 			return nil, err
 		}
 		defer file.Close()
 	}
+
 	return &NodeUsageDb{
 		Path: path,
 		Data: cache.New(365 * 24 * time.Hour, 10*time.Minute),
@@ -35,13 +36,17 @@ func NewNodeUsageDB(path string) (*NodeUsageDb, error) {
 func (m *NodeUsageDb) Load() error {
 	itemsRaw, err := ioutil.ReadFile(m.Path)
 	if err != nil {
-		return errors.Wrap(err, fmt.Sprintln("failed reading file", m.Path))
+		return errors.Wrap(err, fmt.Sprintf("failed reading file %v", m.Path))
+	}
+
+	if len(itemsRaw) == 0 {
+		return nil
 	}
 
 	items := &map[string]cache.Item{}
 	err = json.Unmarshal(itemsRaw, items)
 	if err != nil {
-		return errors.Wrap(err, fmt.Sprintln("failed decoding cache data", string(itemsRaw)))
+		return errors.Wrap(err, fmt.Sprintf("failed decoding cache data %v", string(itemsRaw)))
 	}
 
 	m.Data = cache.NewFrom(365 * 24 * time.Hour, 10*time.Minute, *items)
@@ -52,12 +57,12 @@ func (m *NodeUsageDb) Load() error {
 func (m *NodeUsageDb) Save() error {
 	itemsRaw, err := json.Marshal(m.Data.Items())
 	if err != nil {
-		return errors.Wrap(err, fmt.Sprintln("failed encoding item data", m.Data.Items()))
+		return errors.Wrap(err, fmt.Sprintf("failed encoding item data %v", m.Data.Items()))
 	}
 
 	err = ioutil.WriteFile(m.Path, itemsRaw, 0644)
 	if err != nil {
-		return errors.Wrap(err, fmt.Sprintln("failed wrting file", m.Data.Items()))
+		return errors.Wrap(err, fmt.Sprintf("failed wrting file %v", m.Data.Items()))
 	}
 
 	return nil
